# specs/001-develop-a-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Roblox Trading Platform API
  description: API for secure Roblox item trading with escrow
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /trades:
    get:
      summary: Search trades
      tags: [Trades]
      parameters:
        - name: q
          in: query
          description: Search query for item name
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, NEGOTIATING, PAYMENT_PENDING]
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trade'

    post:
      summary: Create trade listing
      tags: [Trades]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemName, itemImage, price]
              properties:
                itemName:
                  type: string
                  minLength: 1
                  maxLength: 100
                itemImage:
                  type: string
                  format: uri
                description:
                  type: string
                  maxLength: 500
                price:
                  type: number
                  minimum: 0.01
      responses:
        '201':
          description: Trade created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /trades/{tradeId}:
    get:
      summary: Get trade details
      tags: [Trades]
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trade details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '404':
          description: Trade not found

    patch:
      summary: Update trade status
      tags: [Trades]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [NEGOTIATING, CANCELLED]
      responses:
        '200':
          description: Trade updated
        '403':
          description: Forbidden
        '404':
          description: Trade not found
...
  /upload/image:
    post:
      summary: Get signed URL for image upload
      tags: [Uploads]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, fileType]
              properties:
                fileName:
                  type: string
                fileType:
                  type: string
                  enum: [image/jpeg, image/png, image/webp]
      responses:
        '200':
          description: Signed URL for upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                  fileUrl:
                    type: string
        '400':
          description: Invalid file type
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Trade:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        itemName:
          type: string
        itemImage:
          type: string
        description:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [ACTIVE, NEGOTIATING, PAYMENT_PENDING, PAYMENT_COMPLETE, IN_PROGRESS, COMPLETED, CANCELLED, REFUNDED]
        paymentDeadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        creator:
          $ref: '#/components/schemas/User'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/TradeParticipant'
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
    TradeParticipant:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...